More conservative variables, and naming standardisation

+++ libraries/PuLogger/Transmission.h
@@ -21,7 +21,7 @@ struct Transmission {
 	int getRawHum(int sensor) {return humRawValue[sensor];}
   //float getTemp(int sensor) { return float(tempRaw[sensor]) / 10.0;}
   //float getHum(int sensor) { return float(humRaw[sensor]) / 10.0;}
-  bool changed( Transmission other, float tempHys, float humHys) {return (abs(getRawTemp(0) - other.getRawTemp(0)) > tempHys || abs(getRawHum(0) - other.getRawHum(0)) > humHys) || abs(getRawTemp(1) - other.getRawTemp(1)) > tempHys || abs(getRawHum(1) - other.getRawHum(1)) > humHys; } //change to raw values later?
+  bool changed( Transmission other, int tempHys, int humHys) {return (abs(getRawTemp(0) - other.getRawTemp(0)) > tempHys || abs(getRawHum(0) - other.getRawHum(0)) > humHys) || abs(getRawTemp(1) - other.getRawTemp(1)) > tempHys || abs(getRawHum(1) - other.getRawHum(1)) > humHys; } //change to raw values later?
   void printCSV() {
 		printf("%i;%i;%i;%i;\n", int(getRawTemp(0)), int(getRawHum(0)), int(getRawTemp(1)), int(getRawHum(1)));
 	}
